NewUpdate
Time it takes for houses to burn is now dynamic and changes based on quantity of goblins following player. Knight spawn cap now increases based on of quantity of houses burned.

Bugs: After a few houses are burned, the E button stops working when burning down houses. Not sure why this is. Also, a new issue popped up, where when multiple knights kill the same goblin at the same time then the # of goblins in the party gets skewed. Also, knight AI is being finicky and looks like it needs reworked.